[0709] board 부분 database와 연결 완료

+++ src/main/java/com/example/summerspr2025/service/BoardService.java
@@ -7,7 +7,7 @@ import java.util.Map;
 public interface BoardService {
     Map<String, Object> create(Map<String, Object> param);
     Map<String, Object> update(Map<String, Object> param);
-    Map<String, Object> delete(int id);
-    Map<String, Object> detail(int id);
+    Map<String, Object> delete(long id);
+    Map<String, Object> detail(long id);
     Map<String, Object> list();
 }
